reorg ui into modules
